LinearOperator requires dtype specified lobpcg.py
I think that it is a bug in LinearOperator, filed
[skip actions] [skip cirrus]

+++ scipy/sparse/linalg/_eigen/lobpcg/lobpcg.py
@@ -382,7 +382,7 @@ def lobpcg(
     necessary but still supported as the input to `lobpcg`.
     Specifying ``matmat=A_matmat`` explicitely improves performance. 
 
-    >>> A_lo = LinearOperator((n, n), matvec=A_matmat, matmat=A_matmat)
+    >>> A_lo = LinearOperator((n, n), matvec=A_matmat, matmat=A_matmat, dtype=np.int16)
     >>> eigenvalues, _ = lobpcg(A_lo, X, maxiter=80)
     >>> print(eigenvalues)  
     array([100.])
